Updated to OpenFace v2.0.0

+++ libs/GazeAnalyser/include/GazeEstimation.h
@@ -45,10 +45,10 @@ namespace GazeAnalysis
 	void EstimateGaze(const LandmarkDetector::CLNF& clnf_model, cv::Point3f& gaze_absolute, float fx, float fy, float cx, float cy, bool left_eye);
 
 	// Getting the gaze angle in radians with respect to the world coordinates (camera plane), when looking ahead straight at camera plane the gaze angle will be (0,0)
-	cv::Vec2d GetGazeAngle(cv::Point3f& gaze_vector_1, cv::Point3f& gaze_vector_2);
+	cv::Vec2f GetGazeAngle(cv::Point3f& gaze_vector_1, cv::Point3f& gaze_vector_2);
 	
 	// Some utilities
-	cv::Point3f GetPupilPosition(cv::Mat_<double> eyeLdmks3d);
+	cv::Point3f GetPupilPosition(cv::Mat_<float> eyeLdmks3d);
 
 }
 #endif
